added AdditionActivityService to add First and Second Numbers together and store all results in DB
Swagger UI test code 200

+++ Models/Calculations.cs
@@ -5,6 +5,6 @@
         public int Id { get; set; }
         public decimal FirstNumber { get; set; }
         public decimal SecondNumber { get; set; }
-        public int Result { get; set; }
+        public decimal Result { get; set; }
     }
 }
